+++ WebAPI/Controllers/CarsController.cs
@@ -65,7 +65,7 @@ namespace WebAPI.Controllers
         }
 
         [HttpGet("getbydailyprice")]
-        public IActionResult GetByDailyPrice(decimal min, decimal max)
+        public IActionResult GetByDailyPrice(double min, double max)
         {
             var result = _carService.GetByDailyPrice(min,max);
             if (result.Success)
